Fix cast warning.

+++ Library/Source/Nano/Internal/3rd Party/zlib_1.2.3/zutil.c
@@ -303,8 +303,8 @@ voidpf zcalloc (opaque, items, size)
     unsigned size;
 {
     if (opaque) items += size - size; /* make compiler happy */
-    return sizeof(uInt) > 2 ? (voidpf)malloc(items * size) :
-                              (voidpf)calloc(items, size);
+    return sizeof(uInt) > 2 ? (voidpf)malloc((size_t) (items * size)) :
+                              (voidpf)calloc((size_t) items, (size_t) size);
 }
 
 void  zcfree (opaque, ptr)
